+++ docsource/releasenotes.rst
@@ -1,6 +1,21 @@
 Workspace service release notes
 ===============================
 
+VERSION: 0.10.0 (Released 7/2/19)
+---------------------------------
+
+BACKWARDS INCOMPATIBILITES:
+
+* Shock nodes included in workspaces objects as ``@id bytestream`` annotations must now be
+  owned by the user or by the workspace service. If the user owns the node, the node's owner
+  will be changed to the workspace service and write and delete ACLs will be cleared. The
+  node is no longer copied.
+  
+UPDATES:
+
+* Storage systems that support the AWS S3 API can now be used as backend file storage systems
+  as an alternative to Shock or GridFS.
+
 VERSION: 0.9.1 (Released 4/18/19)
 ---------------------------------